simple_blog/requirements.txt
asgiref==3.7.2
Django==5.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.8,<4.0
djangorestframework-simplejwt
